+++ core/src/main/java/com/populaire/projetguerrefroide/dao/builder/RegionStoreBuilder.java
@@ -1,6 +1,7 @@
 package com.populaire.projetguerrefroide.dao.builder;
 
 import com.github.tommyettinger.ds.IntList;
+import com.github.tommyettinger.ds.LongList;
 import com.github.tommyettinger.ds.ObjectIntMap;
 import com.populaire.projetguerrefroide.map.RegionStore;
 
@@ -8,7 +9,7 @@ public class RegionStoreBuilder {
     private final int defaultCapacity;
     private int index;
     private final ObjectIntMap<String> regionIds;
-    private final IntList buildingIds;
+    private final LongList buildingIds;
     private final IntList buildingValues;
     private final IntList buildingStarts;
     private final IntList buildingCounts;
@@ -17,7 +18,7 @@ public class RegionStoreBuilder {
         this.defaultCapacity = 396;
         this.index = 0;
         this.regionIds = new ObjectIntMap<>(this.defaultCapacity);
-        this.buildingIds = new IntList(this.defaultCapacity);
+        this.buildingIds = new LongList(this.defaultCapacity);
         this.buildingValues = new IntList(this.defaultCapacity);
         this.buildingStarts = new IntList(this.defaultCapacity);
         this.buildingCounts = new IntList(this.defaultCapacity);
@@ -39,7 +40,7 @@ public class RegionStoreBuilder {
         return this;
     }
 
-    public RegionStoreBuilder addBuilding(int buildingId, int value) {
+    public RegionStoreBuilder addBuilding(long buildingId, int value) {
         int startIndex = this.buildingStarts.get(this.index);
         int endIndex = startIndex + this.buildingCounts.get(this.index);
         for(int i = startIndex; i < endIndex; i++) {
